new dashboard app auth
