+++ zakat/dashboard/auth/forms.py
@@ -1,0 +1,6 @@
+from django import forms
+
+
+class LoginForm(forms.Form):
+    username = forms.CharField()
+    password = forms.CharField(widget=forms.PasswordInput)
