+++ zakat/dashboard/auth/views.py
@@ -1,0 +1,36 @@
+from django.contrib import messages
+from django.shortcuts import render
+from django.contrib.auth import authenticate, login as auth_login
+# Create your views here.
+
+
+# TODO redirect to the 'Gatherings' page in the future
+# @user_passes_test(lambda user: user.is_anonymous,
+#                   login_url='http://127.0.0.1:8000',
+#                   redirect_field_name=None)
+from dashboard.auth.forms import LoginForm
+from dashboard.auth.helper import check_is_employee
+
+
+def login(request):
+    """Simple login page for employees only"""
+    if request.method == 'POST':
+        form = LoginForm(request.POST)
+        if form.is_valid():
+            cleaned_data = form.cleaned_data
+            user = authenticate(username=cleaned_data['username'],
+                                password=cleaned_data['password'])
+            if user is not None:
+                if not check_is_employee(user):
+                    messages.info(request, f'User {user.username} is not an employee!')
+                else:
+                    auth_login(request, user)
+                    messages.info(request, 'Authenticated successfully!')
+            else:
+                messages.error(request, 'User credits are wrong!')
+        else:
+            messages.error(request, 'Given data is not valid!')
+    else:
+        form = LoginForm()
+
+    return render(request, 'dashboard/auth/login.html', {'form': form})
