+++ zakat/dashboard/forms.py
@@ -21,9 +21,6 @@ class EmployeeCreateForm(forms.ModelForm):
             employee.save()
 
 
-class LoginForm(forms.Form):
-    username = forms.CharField()
-    password = forms.CharField(widget=forms.PasswordInput)
 
 
 class PostCreateForm(forms.ModelForm):
