+++ zakat/dashboard/urls.py
@@ -1,7 +1,5 @@
 from django.urls import path, include
 from .views import *
-from dashboard.views import login
-
 app_name = 'dashboard'
 
 urlpatterns = [
@@ -13,7 +11,6 @@ urlpatterns = [
     path('staffs/delete/<int:pk>/', EmployeeDelete.as_view(), name='employee_delete'),
     # Here you write all your paths
     # For example 'staffs/', 'staffs/create'
-    path('login/', login, name='login'),
     path('partners/', PartnerList.as_view(), name="partners_list"),
     path('partners/create/', PartnerCreate.as_view(), name="partner_create"),
     path('partners/<int:pk>/', PartnerDetail.as_view(), name="partner_detail"),
